Web app acceptable
The web application runs as it should. A user visits the page, and
inserts information. Inputted information is displayed on the next page
with a link to the homepage. Once the user goes back to the homepage he
or she can see their previous entry.

Future Improvements:
The application could be better designed. If a user does not select any
interests, then when interests are listed there should be a message that
says none.